feat: Add support for checking all changes of PR
This Adds support for checking entire PR context instead of only the last commit.
Also fixes a bunch of errors making sure it uses the correct Repo when looping over the list. When setting `RepoConfig` it Assigned the object to the last one in the loop of Repositories. Resulting in wrong comparision.